moved bemio import function to separate file
